Made stopStream optional on stop
For the same reason that I needed the stream to be passed on init, I don't want the stream to stop when I stop recording. 
I continuously show the stream, independently of recording.

You don't have to add this if you don't want to, but I like to use it this way. 

I didn't have this problem before, because the stream didn't actually stop earlier.